requirements.txt
pyarrow>=11.0.0
redis>=5.0.1
pandas>=1.1.3
python-dateutil>=2.8.1